requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
Werkzeug
openai-whisper
imageio-ffmpeg
transformers
torch
sentencepiece
